The lcd and dsi sytem loads. Removing the useless data-anes attributes.
But the driver fails initialisation. Somewhere during my test, I fried
the panel or the dsi output because the vendor normal boot shows no
image anymore.
Therefore, I will replace the A33 on the board. And on the bricked one,
replace the emmc with a new one programmed using a usb reader for BGA153
EMCs. And see how it behaves.